+++ response.go
@@ -14,37 +14,37 @@ type Response interface {
 }
 
 type response struct {
-	writer      http.ResponseWriter
+	Writer      http.ResponseWriter
 	Status      int
 	Size        int64
-	isCommitted bool
+	IsCommitted bool
 }
 
 func NewResponse(w http.ResponseWriter) Response {
-	return &response{writer: w}
+	return &response{Writer: w}
 }
 
 func (r *response) WriteHeader(code int) {
-	if r.isCommitted {
+	if r.IsCommitted {
 		fmt.Println("response has already committed")
 		return
 	}
 
 	r.Status = code
-	r.writer.WriteHeader(r.Status)
-	r.isCommitted = true
+	r.Writer.WriteHeader(r.Status)
+	r.IsCommitted = true
 }
 
 func (r *response) Write(b []byte) (int, error) {
-	if !r.isCommitted {
+	if !r.IsCommitted {
 		if r.Status == 0 {
 			r.SetStatus(http.StatusOK)
 		}
 		r.WriteHeader(r.Status)
-		r.isCommitted = true
+		r.IsCommitted = true
 	}
 
-	n, err := r.writer.Write(b)
+	n, err := r.Writer.Write(b)
 	r.Size += int64(n)
 
 	return n, err
@@ -55,5 +55,5 @@ func (r *response) SetStatus(code int) {
 }
 
 func (r *response) Header() http.Header {
-	return r.writer.Header()
+	return r.Writer.Header()
 }