+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 0.16.1
+# 0.x.x
+
+## 0.17.x
+
+### 0.17.0
+
+- timeout middleware
+- poteto.Response.writer became public member
+
+## 0.16.x
+
+### 0.16.1
 
 - fix bug
   - become: `Context.QueryParam()` & `Context.PathParam()` only return string